Add custom exception handling

+++ src/main/java/com/payconiq/tradingappws/service/StockServiceImpl.java
@@ -2,10 +2,18 @@ package com.payconiq.tradingappws.service;
 
 import com.payconiq.tradingappws.dao.entity.Stock;
 import com.payconiq.tradingappws.dao.repository.StockRepository;
+import com.payconiq.tradingappws.dto.model.StockCreateDto;
+import com.payconiq.tradingappws.dto.model.StockQueryDto;
+import com.payconiq.tradingappws.dto.model.StockUpdateDto;
+import com.payconiq.tradingappws.exception.DuplicateStockException;
+import com.payconiq.tradingappws.exception.StockLockedException;
+import com.payconiq.tradingappws.exception.StockNotfoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.modelmapper.ModelMapper;
 
 import java.util.*;
+import java.util.stream.Collectors;
 
 @Service
 public class StockServiceImpl implements StockService{
@@ -13,24 +21,37 @@ public class StockServiceImpl implements StockService{
     @Autowired
     private StockRepository stockRepository;
     
+    @Autowired
+    private ModelMapper modelMapper;
+    
     @Override
-    public List<Stock> getAllStocks() {
-        return stockRepository.findAll();
+    public Set<StockQueryDto> getAllStocks() {
+        return stockRepository.findAll()
+                .stream()
+                .map(stock -> modelMapper.map(stock, StockQueryDto.class))
+                .collect(Collectors.toCollection(TreeSet::new));
     }
     
     @Override
-    public Stock createStock(Stock stock) {
-        int uniqueId = generateUniqueId();
-        if (uniqueId == -1){
-            return null;
+    public StockQueryDto createStock(StockCreateDto stockCreateDto) {
+        // Verify that stock with the same name does not exist
+        Optional<Stock> duplicateStock = Optional.ofNullable(stockRepository.findByName(stockCreateDto.getName()));
+        if (!duplicateStock.isPresent()) {
+            int uniqueId = generateUniqueId();
+            if (uniqueId == -1){
+                return null;
+            }
+            Stock stockModel = new Stock()
+                                       .setId(uniqueId)
+                                       .setName(stockCreateDto.getName())
+                                       .setCurrentPrice(stockCreateDto.getCurrentPrice())
+                                       .setCreationDate(new Date())
+                                       .setLocked(true);
+            stockRepository.save(stockModel);
+            return modelMapper.map(stockModel, StockQueryDto.class);
         }
-        Stock newStock = new Stock();
-        newStock.setId(generateUniqueId());
-        newStock.setName(stock.getName());
-        newStock.setCurrentPrice(stock.getCurrentPrice());
-        newStock.setCreationDate(new Date());
-        newStock.setLocked(true);
-        return stockRepository.save(newStock);
+        String message = "Stock with name : "+ stockCreateDto.getName() +" already exists.";
+        throw new DuplicateStockException(message);
     }
     
     private int generateUniqueId() {
@@ -49,33 +70,50 @@ public class StockServiceImpl implements StockService{
     }
     
     @Override
-    public Stock getStockById(int id) {
+    public StockQueryDto getStockById(int id) {
         Optional<Stock> stock = stockRepository.findById(id);
-        return stock.orElse(null);
+        if (stock.isPresent()){
+            return modelMapper.map(stock.get(), StockQueryDto.class); 
+        }
+        String message = "Stock with id : "+ id +" not found";
+        throw new StockNotfoundException(message);
     }
     
     @Override
-    public Stock updateStockPrice(int id, Stock stockUpdateDto) {
+    public StockQueryDto updateStockPrice(int id, StockUpdateDto stockUpdateDto) {
         Optional<Stock> stock = stockRepository.findById(id);
         if (stock.isPresent()) {
-            Stock stk = stock.get();
-            stk.setLastUpdate(new Date());
-            stk.setLocked(true);
-            stk.setCurrentPrice(stockUpdateDto.getCurrentPrice());
-            stk = stockRepository.save(stk);
-            return stk;
+            Stock stockModel = stock.get();
+            if(!stockModel.isLocked()){
+                stockModel.setLastUpdate(new Date());
+                stockModel.setLocked(true);
+                stockModel.setCurrentPrice(stockUpdateDto.getCurrentPrice());
+                stockModel = stockRepository.save(stockModel);
+                return modelMapper.map(stockModel, StockQueryDto.class);
+            } else {
+                String message = "Stock with id : "+ id +" is locked. Please try after sometime.";
+                throw new StockLockedException(message);
+            }
         } else {
-            return null;
+            String message = "Stock with id : "+ id +"  not found";
+            throw new StockNotfoundException(message);
         }
     }
     
     @Override
-    public Stock deleteStock(int id) {
+    public StockQueryDto deleteStock(int id) {
         Optional<Stock> stock = stockRepository.findById(id);
         if(stock.isPresent()){
-            stockRepository.deleteById(id);
-            return stock.get();
+            Stock stockModel = stock.get();
+            if(!stockModel.isLocked()){
+                stockRepository.deleteById(id);
+                return modelMapper.map(stock.get(), StockQueryDto.class);
+            }
+            String message = "Stock with id : "+ id +" is locked. Please try after sometime.";
+            throw new StockLockedException(message);
         }
-        return null;
+        String message = "Stock with id : "+ id +"  not found";
+        throw new StockNotfoundException(message);
     }
+    
 }
